Add Vector support to Register Assigner on x
Register Assigner on x now recognizes "TR_VRF" virtual register as
being a vector register and will attempt to assign a XMM register to
them.

Created method isFPRUsedAsVFR. It checks if the processor supports SSE2
and if so attempts to assign vector registers if the system has TR_FPR
real registers.

Created method getSupportsOpCodeForAutoSIMD. It is used to check which
vector opcode are currently supported for autoSIMD. Currently there are
none so it always returns false.

Added code to handle spilling and filling of registers for 128 bit
vector sized data.

Issue: #431
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>