Stop inlining MethodHandles in fsd mode
We need to stop inlining MethodHandle in fsd because those trees are
generated by JIT out of thin air and dont have corresponding
bytecode for interpreter to transit from. Note that this changeset
only solves part of the problem and there are still trees generated
from ilgen without inlining that will be fixed later.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>