WIP: Safely dup trees for versioning wrtbar removal
When creating guard tests outside loops in order to remove wrtbars,
LoopVersioner creates the test expression by duplicating
other expressions from elsewhere.

In order to do this safely LoopVersioner needs to make sure to reset
node flags for the duplicated trees because they may not be valid at
the location where the test is performed. If these flags are not reset
later optimizations may perform incorrect transformations based on
the flags.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>