clean up singleImplementer related code
Separate the concept of SingleImplementer and SingleJittedImplementer
when looking for callsite target to inline. SingleImplementer is the only
target that can be called from a certain callsite and is reliable.
SingleJittedImplementer is more of a guess that the possibility of calling
a JITted target from a callsite should be greater than the interpreted ones.
Therefore SingleJittedImplementer should always be the last resort.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>