Relax CFGEdgeList implementation dependencies
Currently there are many locations where we make the assumption that
TR::CFGEdgeList is implemented using std::list. The most prolific
assumption is assuming the availability of a remove() function.

This change set eliminates those dependancies enabling us to use any of
the standard sequence containers.

**NOTE: This change does not attempt to address or in any way rectify
iterator and reference invalidation safety.  Attempting to use a
container with stricter requirements would require a thorough review and
updating of our container usage.**

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>