Find stores/returns after inlined calls in OSR
In the event that the result of an inlined call is referenced once
after it, in a store or a return, the inliner is able to use that
treetop directly rather than requiring a temp to return the result.

This behaviour is broken by postExecutionOSR as all calls that are
OSR points will be followed by a pending push that stores its
result. The correction here skips this pending push to attempt
the existing analysis. This is valid under OSR as the inlined call
is no longer an OSR point and its result should not be on the stack
if it is stored or returned immediately.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>