Enable folding for test of array class in VP
Detect Iand operating on a fixed array class so that optimizations
can fold away runtime tests for array types. This extends the
existing functionality in VP for folding an isArray test based on
an object to the case where we don't have an object, but we do have
a Class in hand.

Signed-off-by: JamesKingdon <jkingdon@ca.ibm.com>