Use non-offset BCI in supportsInduceOSR check
induceOSRAfter will perform a supportsInduceOSR check to verify the
provided bytecode info is a potential OSR point. However,
induceOSRAfter may be provided with an offset bytecode index under
postExecutionOSR, whilst supportsInduceOSR always expects the original
index. This change allows the actual BCI and an offset to be
specified seperately to induceOSRAfter so that it performs the
correct check.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>