Handle method entry for involuntaryOSR
In involuntaryOSR, method entries, although not explicitly represented
in the trees, are yield points. Therefore OSRCatchBlock is not optional
for this mode. This change adds asynccheck at method entry to represent
the OSR point if there is no other explicit OSR points in this method.
Otherwise, the rest of the optimizer would try to remove the OSR infrastructure.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>