Add option to force induceOSRAfter always
induceOSRAfter will check if the provided bytecode index
supports induce OSR before inserting the OSR infrastructure.
However, there are times where it is known to the caller that
a transition must be possible, for example, at the start of
the method. This adds an option forcing the method to skip
its safety check.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>