[CMake] Start creating toolconfig
A toolconfig is a file that will export tool-dependent definitions for
use in other contxts.

This commit moves the toolchain dependent definition of the warnings as
error flag to fragment files named after the toolchain, as well as
adding a toolchain validator to ensure that the toolchain has exported
all the required definitions.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>