Do not use OMR_GLUE as a subdirectory
We can't realistically use the OMR_GLUE as a subdirectory. This will
break when the glue is outside OMR. Instead, we have to rely on a parent
project defining all the right glue targets and variables, before
building omr.

Instead of testing for the existence of the OMR_GLUE variable,
The OMR_EXAMPLE flag can be used to conditionally add the example
directory, which will define the glue library.

OMR_EXAMPLE is already enabled in the CI.

There is an issue of interdependence here that we have yet to sort out.

Signed-off-by: Robert Young <rwy0717@gmail.com>