Change TOC size from 256 to 64 in -Xquickstart mode
We are changing the TOC size from 256KB back to the original 64KB in -Xquickstart mode in an effort to decrease footprint.

The default TOC size was changed from 64KB to 256KB in commit bdc1efa to help with performance in Liberty/Hadoop/Spark. However, since throughput is not very important in -Xquickstart mode, we can afford to revert back to 64KB in this mode.

This change will decrease footprint.

Signed-off-by: Harry Yu <harryyu1994@gmail.com>