Prepare to relocate ValuePropagation functions to OpenJ9
Functions that are wholly guarded with J9_PROJECT_SPECIFIC macros will
be relocated to the OpenJ9 project.  The code that calls them, however,
is guarded with J9_PROJECT_SPECIFIC macros but is nested within functions
that are not.  To enable relocation, these calls need to have their
receiver object (i.e., `this`) downcast to `TR::ValuePropagation`.

Functions affected:

```
transformRTMultiLeafArrayCopy
transformObjectCloneCall
transformArrayCloneCall
transformConverterCall
```

Signed-off-by: Daryl Maier <maier@ca.ibm.com>