Silence ignored-return warnings
This is a minimal change to remove the warnings when building. A
proper fix requires a deeper redesign of the code as not all of
these functions return a value and the callers aren't prepared
to handle the failure / error cases.

Silencing the warning seems like a pragmatic step towards less
noise in the builds.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>