Remove .gitattributes file as it breaks travis builds
Travis builds have been updated to newer versions of git which recognize
the worktree-encoding attribute. This means the current files are
invalid as they force the encoding of all files as ebcdic.

Other builds, and developers, will be broken when they update to
newer versions of git.

This solution won't work long term. Removing the file now so we
can unblock builds while thinking about the right long term solution.

This backs out part of #2568

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>