Specialize Barrier for Value slots
Barriers assume that you are overwriting a reference slot with another
reference. In the case of Values, the previously overwritten value may
not be a reference. The Value specialization checks that the previous
value was a reference before pre-barrier-ing.

The slot handle parameter was moved to the class to permit partial
specialization.

Signed-off-by: Robert Young <rwy0717@gmail.com>