Final Commit for Copy Scan Work Experimentation
In summary, by changing copy scan caches to pure copy caches would
result in footprint savings that would be beneficial for pure BF
scanning, but not so for hierarchical scanning. Even for BF it would not
be much. Overall, it's just not worth it making the code more complex.
This work may be revisited for future work.