WIP: Incorporate Intel Transactional Memory (TSX) into OpenJ9 Locking
Transactional Lock Elision

Currently, the J9JIT (with TLE enabled) transforms monenters/monexits
of synchronized blocks to execute transactionally, if analyisis of
the block passes certain conditions.

The issue arises when some synchronized blocks are transformed to
use TLE, while others are not, AND they are synchronized on the same
lock object.

A thread cannot commit a transaction if the lock is also synchronized
by another thread by non-TLE locking. This creates excessive aborts.

This patch is a work-in-progress implementation of Restricted
Transactional Memory (RTM) as applied to OpenJ9 locks and is intended
to replace the TLE implementation in J9JIT. This approach avoids the
aformentioned problems by eliding locks at the lock object level as
opposed to the lock sites.

See https://github.com/eclipse/openj9/issues/2484

Signed-off-by: John Xu <xujohn@ca.ibm.com>