Fix for incorrectly marked collected reference on X86
The result register of TR::aiadd/TR::aladd was incorrectly marked as a
collected reference, when the aiadd/aladd node is not internal pointer.
As a result, an invalid GC map entry may be introduced and causes GC assertions.
The result of TR::aiadd/TR::aladd can never be a collected reference, fixing it.

Signed-off-by: Victor Ding <dvictor@ca.ibm.com>