Provide an API to get the constant values from an ILValue
I have been working on an InterpreterBuilder class for generating
an Interpreter using JitBuilder. One of the main goals is to allow
the BytecodeBuilder objects used by the InterpreterBuilder to be
consumed without any changes to generate a JIT method. When using
the VirtualMachineState hierarchy some operations require getting
the constant value out of an ILValue so that calls to VMOS Push
and Pop can index into its array of IlValues.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>