Dynamically adjust TLH maximum size & support allocation sampling
The goal of this PR is to force out-of-line allocation periodically
according to sampling interval requested such that downstream projects
can report such events.
The approach here is to adjust TLH maximum size dynamically to match
sampling interval (when it is larger than minimum TLH size) or actual
TLH maximum size whichever less. When a TLH refresh is performed,
current TLH allocation size is recorded, and force out-of-line
allocation if the sampling interval has been reached.

Adding a new field GCExtensionsBase.tlhActiveMaximumSize as current
maximum size for TLH allocation, could be changed as per
allocationSamplingInterval (new field as well). By default, it is same
as tlhMaximumSize.

Also adding fields enableAllocationSampling, allocationSamplingInterval
and currentAllocationRemainder to support allocation sampling;

MM_TLHAllocationSupport::refresh() is modified to record the allocation
size of the TLH being refreshed. In addition, disableInlineTLHAllocate()
to force out-of-line for next allocation when the sampling interval has
been reached;

Adding a few trace entries.

Note: forcing out-of-line allocation allows downstream project
https://github.com/eclipse/openj9 to report sampling events according to
the interval requested.

Signed-off-by: Jason Feng <fengj@ca.ibm.com>