Add isinstance check to __eq__ overloads in GenUtils classes
Adding a call to `isinstance` ensures that other, internal equality
checks are only performed if the "other" object is an instance of the
current class. Otherwise, if "other" is of some other type (e.g. None),
trying to compare the fields of the objects may result in an exception
if the objection doesn't have those fields.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>