Split the example glue into base and extended
When the OMR GC API is enabled, certain elements of the glue will be
provided by the GC. To ensure that every component finds the correct
copy of the glue headers, the example's version has to be taken out of
the include path.

To facilitate this, the example's glue headers are split into an include
and an include_extended directory. The include_extended directory is
only added to the includes when OMR_GC_API is OFF.

As well, some GC glue sources are only added if the GC API is off, since
those are provided by rthe GC as well.

Signed-off-by: Robert Young <rwy0717@gmail.com>