Enable aliasing of auto variables that have their address taken
The _hasCallsAsUses flag in TR_UseDefInfo controls whether function
calls will be considered as having aliased local symbols. This flag is
set to false in OMR as for Java like languages function calls cannot
alias local symbols. However for a C like language where the address of
a local variable may be passed to the called function, it is possible
for the function call to alias certain local symbols. This parameter
needs to be made configurable, i.e. there needs to be an API to modify
it. For now, this is being changed to true here to flag it as a
requirement for this set of changes.

Signed-off-by: Dibyendu Majumdar <mobile@majumdar.org.uk>