Create a C API for OMR low-level IL interface
While there is a higher level JIT interface in the form of JITBuilder
api for which a C api is being worked on, having a C api for the lower
level IL interface is useful. In particular, the C api that is part of
this pull request was used to implement JIT backends for C as well as
Ravi, a derivative of Lua 5.3.

Signed-off-by: Dibyendu Majumdar <mobile@majumdar.org.uk>