Fix bugs in Type-Coercion Evaluator
When normalizing values in the type-coercion evaluator
on Z, we need to shift a 32-bit integer left into the
high word. This is required to ensure the Test Data class
floating-point instruction behaves correctly. Additionally,
logic for generating literal pool values and then loading
it in a register is removed. This work can be performed by
the generateRegLitRefInstruction that picks the best method
to load a constant.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>