Use constraints to remove bounds check for table switch
Currently only lookup switches are handled by VP. This change
allows the bounds check to be removed from a table switch if all
of the cases are covered based on the selector type.

Also in SwitchAnalyzer propagate the safeToSkipBoundsCheck similar
to cannotOverflow since it is used by some platforms instead of
cannotOverflow.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>