Added option for parameters in various OMR build scripts
If a specific repo, branch or label is provided, then the
script will use it. Otherwise these values default to the
values that already existed in the script. Additionally
the script will check if ccache is installed on the machine
before running the ccache command, so the job won't fail
on machines that don't have ccache (e.g. docker containers).
[skip ci]

Signed-off-by: Colton Mills <millscolt3@gmail.com>