Use atomic exchange to install nodes to the MCS lock queue
Previously, the compare and swap (CAS) operation was used to install a
node to the MCS lock queue. If the CAS fails, then the MCS algorithm
malfunctions.

Now, the atomic exchange operation is used to install a node to the MCS
lock queue. In an atomic exchange operation, the swap always happens.
The MCS algorithm works fine with this change.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>