Address hangs
The algorithm spins on MCS locks. If the MCS lock is not acquired
while spinning, it blocks using the thread's condition variable and
monitor's mutex. Transition from spinning to blocking state leads to
hangs where the blocked thread does not get notified or woken up.

The following steps resolve the hangs:
1) Acquire MONITOR_LOCK before omrthread_mcs_unlock.
2) Check mcsNode->blocked before blocking on monitor mutex.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>