Add an alignment nop instruction class on Power
In some cases, it is beneficial to ensure that an instruction is aligned
to a specific number of bytes by padding the instruction stream with
nops. The Power codegen was previously doing this ad-hoc in the binary
encoder, but a better solution would be to allow tree evaluators to
signal where this should happen. This is now possible using the new
PPCAlignmentNopInstruction class of instruction. When binary encoded,
this instruction will generate the necessary number of nops to ensure
that the next instruction is aligned on the requested boundary.

Signed-off-by: Ben Thomas <ben@benthomas.ca>