Make group-ending nops apparent in the tracefile
Previously, the genop instruction had the unhelpful name of "nop",
making it indistinguishable in tracefiles from the normal nop
instruction. This caused some confusion when debugging the new loop
label alignment, as it made it seem like the group-ending nop was not
being emitted. In order to make the difference apparent, group-ending
nops will now show up as genop in the tracefile.

Signed-off-by: Ben Thomas <ben@benthomas.ca>