JitBuilder2 (JB2) prototype
First draft prototype of an enhancement to the JitBuilder API to
evolve it into an intermediate representation (IR) that can be
analyzed and transformed ahead of being converted to native
instructions using the OMR compiler. More documentation will
follow, but this commit includes the prototype implementation
located, for the moment, in the jitbuilder/jbil directory.

The use of the prototype is demonstrated by a slightly modified
version of the matrix multiply code sample along with a few other
smaller samples that I've used to do some testing. The demonstration
uses almost the same raw code sample but showcases some of what
can be done with the prototype.

A TextWriter class can generate a readable log of the current
state of the JB2 representation. The TextWriter class extends a
Visitor class that uses a worklist to traverse all the
builder objects and all the operations in a function.

Another such class is Transformer, which allows builders
and operations to be replaced as they are visited. Three
Transformer subclasses are provided: an extremely simple
AppendBuilderInliner that replaces an AppendBuilder operation
with the builder it appends if that builder is not the
target of any branching operation, a DialectReducer that
at the moment just converts a ForLoop operation into lower
level conditional branches, and a CodeGenerator that calls
through the original JitBuilder API to the OMR compiler to
generate native code for a function. The CodeGenerator is
not yet complete but supports most of the kinds of operations
one can create with this prototype (LoadField being an
example of one that *isn't* yet supported.

One of the more interesting classes is Debugger, which
provides an IL level debugger for a FunctionBuilder.
Rather than compiling a FunctionBuilder (which generates
a native function entry point to call and execute the
function), you can ask for a DebugEntry which also
returns a native function entry point that, when called,
instantiates a Debugger which then pauses at the beginning
of the function's IL. The debugger (jbdb) has a gdb feel
to it and supports single stepping, setting of breakpoints
(before/after operations, builders, and at specific time step),
printing of values and symbols.  The Debugger implementation
is also interesting for a couple of reasons. First, it uses
a dynamically generated structure (that contains fields of
every kind of Type referenced by the original FunctionBuilder,
including user types, pointer types, etc.). Second, the
Debugger automatically compiles each operation individually
to support single stepping that operation. This compilation
process automatically transforms all the operands, symbols,
etc. that are reference by the IL operation into the
corresponding values maintained by the debugger as it
executes the target function. The Debugger even supports
user defined types (I tested it with a Complex type) without
any modification thanks to another class (see below) called
TypeReplacer. The Debugger is actually a quite complicated
compiler in its own right that demonstrates some of the power
of JitBuilder2, essentially implementing an interpreter
for JitBuilder2 IL without requiring a separate implementation
of the semantics of those IL operations. The exact same
Debugger implementation works both with the included matrix
multiply (which operates on Doubles) as well as with an example
I've created (but it's currently hard to include in the PR)
that operates on a user Complex class.

A key enabler for the Debugger is a class called TypeReplacer.
This subclass of Transformer translates operations and types
automatically according to how it's configured. TypeReplacer
can either replace references (and values) of one Type to
another Type, or it can "explode" a Type. Exploding a Type
is how user defined types are handled generically. You can
create a Type object that provides a "layout" type (which
must be a StructType that describes the shape of the Type
when "exploded") and also an "explode" function that can
break apart a LiteralValue of the user class into its
constituent LiteralValues (for each of the fields of the
layout type). TypeReplacer maintains a set of maps between
the old types and the new types, and can clone operations
multiple times to deal with exploded types. It has been
verified for a relatively simple Complex class and only
really required special handling for multiplication and
any Complex specific operations (e.g. Conjugate, Magnitude,
etc.)..

Creating new Types and Operations in this prototype
requires modifying the source files to add support
for them. A future enables this extension to be done
in a more programmatic way (but slightly more clumsy from
a user perspective).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>