No freeSpill when doing local RA for ICF
Currently RA records the register states for mainline code when starts
assigning registers to out-of-line instructions so that out-of-line
instructions respects the register assignments in the mainline code.
This change let RA remember the state of spills as well by blocking
freeSpill because out-of-line instructions should have the same spills
as mainline code as well.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>