Update double mapping decision to share memory
In current implementation of double mapping, heap is
is create as shared and with a file descriptor associated
with it. Instead of sharing the heap we’ll only share the
region required by the double mapped arraylet. That
way we save shared memory, meaning, other objects
that do not rely on double mapping will not contribute
to shared memory allocated. That’s good because Linux
systems have a limit on the amount of shared memory
that can be used at a given time; the default is half of
physical memory. However, this new approach introduces
the a new limitation. There is a limit in the amount of file
Descriptors that can be used at a given time; systems
normally default a soft limit of 1024 file descriptors per
process.

Fixes: https://github.com/eclipse/openj9/issues/8621

Signed-off-by: Igor Braga <higorb1@gmail.com>