Change node API for monitor of value type
The hidden child of a monitor related node is used to store the
monitor object's type information. The downstream projects can
generate more optimized code accordingly.

The type information has the following two forms:

1. The persistent class identifier is stored if the type of a
monitor object is a known class. Notice java.lang.Object identifier is only
stored if it's a fixed class

2. IDENTITY_CLASS_PLACEHOLDER is stored if the exact
monitor type is unknown but is of identity class type

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>