Improve the code for handling of memory stats from cgroup
There are multiple issues in the code that gather memory stats
by combining the results from the cgroup and the host.
1. If file memory.memsw.limit_in_bytes is not present, then the code
incorrectly sets swap limit to 0 (by setting
OMRCgroupMemoryInfo.memoryAndSwapLimit to same value as
OMRCgroupMemoryInfo.memoryLimit). Absence of memory.memsw.limit_in_bytes
actually indicates swap support or swap accounting is not configured
in the kernel which implies there is no limit on swap usage by the
process of that cgroup.
2. If memory.memsw.limit_in_bytes is not present, then
memory.memsw.usage_in_bytes will also not be there, and thus the
call to read memory.memsw.usage_in_bytes can be avoided.
3. The logic in retrieveLinuxMemoryStats() that combines the memory
stats of the cgroup with that of the host was unnecessarily complex.
Eg there is no scenario where OMRCgroupMemoryInfo.memoryUsage can be
more than OMRCgroupMemoryInfo.memoryLimit. This patch removes such
unnecessary checks.
4. The code in retrieveLinuxMemoryStats() was returning host stats if
OMRCgroupMemoryInfo.memoryLimit is more than the system memory.
This is fine if the memoryLimit is not set, in which case it gets a huge
value close to LONG_MAX. However in an unlikely case of cgroup memory limit
set to more than host memory, it is still possible to get other memory
stats like swap information from the cgroup.
5. Added a new global variable PPG_cgroupMemoryParamNotSetValue that
stores the value of the memory (or memory+swap) limit when it is not set.
If a particular limit is not set, then it gets a default value which is
LONG_MAX round down to page size. Since getting page size involves a
system call, it is better to cache this value in a global variable.
6. Fixed the format specifier used for reading the cgroup stats from the
files.

Signed-off-by: Ashutosh Mehra <mehra.ashutosh@ibm.com>