Temporarily disable SI(16) range assert on Power
Previously, instructions filling in a 16-bit signed SI field in the
Power codegen would check that the immediate was within the expected
range of -0x8000 to 0x7fff. Recently, several new issues were found with
out-of-range immediates being used in OpenJ9 and there's a worry that
there may be more lurking in the code. Since the OpenJ9 0.21 release is
quite close, this assert is being temporarily disabled to avoid
triggering in this release.

Signed-off-by: gita-omr <koblents@ca.ibm.com>