OMR AOT: Relocation Record files addition
The added files are: OMRRelocationRecord.hpp, OMRRelocationRecord.cpp.
These files contain RelocationRecord, RelocationRecordGroup,
RelocationRecordBinaryTemplate classes. The RelocationRecordGroup is an
iterator for application of relocations. The
RelocationRecordBinaryTemplate is the class that has RelocationRecord
information. Those classes realize core relocations storage functiona-
lity in OpenJ9, but are language independent and thus are worth inclu-
ding into OMR.
Compared to OpenJ9, the RelocationRecord classes became extensible.
Another proposed change is in the creation of relocation records:
the implementation of TR::RelocationRecord::create function is proposed
to be done as a function in corresponding relocationRecord class. The
RelocationRecordBinaryTemplate, has undergone changes to improve
readability and portability: the setSize() and setType(), as well as
x64 architecture specific setExtra() and their corresponding accessor
methods are now  RelocationTarget API calls, as advised in Issue #3031

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>