Revert "Fix lbyteswap store optimization on ppc32"
As discussed in eclipse/omr#5717, this fix uses a temp memory
reference that is not used in an instr. When the symbol is
unresolved the MemRef constructor creates an UnresolvedDataSnippet
that does not have an address in mainline code to return to.
Which causes an assert to fail. Fix is safe to remove as the
reverseStore path is temporarily disabled (eclipse/omr#5685)

This reverts commit 9b282a1136f078ff0d04b5719948b06137628d48.

Signed-off-by: Abdulrahman Alattas <rmnattas@gmail.com>