Relax the assertion in RegDepCopyRemoval
While replacing with new node in RegDepCopyRemoval, we check if the
passthrough node attached to the register dependency has corresponding
regStore which is needed if if later optimization decides to split the
blocks. Currently we were throwing assert while creating
copyToNewVirtualRegister node under dependency where a global register
used by passThrough is different than the global register used by the
value node under it. As we are going to generate a regStore for that
passthrough node, we do not need this strict assert.

Fixes: eclipse/openj9#11943

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>