Increase size of buffer in block validation
Previously, a buffer size of 100 characters was passed to sprintf in the
block validator. However, the format string itself is 94 characters,
which doesn't leave enough room in the buffer for the print arguments.
As a result, running the block validator (e.g. with paranoidOptCheck)
can lead to a buffer overflow.

This commit fixes the issue by increasing the size of the buffer to 150
characters.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>