Add comments to isArrayCompTypeValueType
Added comments to clearly document the cases considered in
isArrayCompTypeValueType.

Also corrected handling of any array whose signature is known, and is
either a fixed class or whose signature does not match
"[Ljava/lang/Object;".  It was being treated as not an array of a value
type.  Instead, it should only be considered as not an array of value type
if it is a fixed class whose signature matches "[Ljava/lang/Object;".

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>