Cleanup
We don't need to check `isInternalPointer` before creating new pinning
array pointers because we wouldn't have made it that far if
`isInternalPointer` was false. Same for when incrementing internal
pointer count.

We can check the opcode of second child in the outer if because we don't
do anything if the second child opcode isn't one of those two. Similar
to previous if block, we can check `isInternalPointer` in the outer
most if block because we don't do anything if the bool if set to false.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>