Normally a TLH can span multiple sweep chunks without any issues. However with SATB, we require the TLH to be contained within a single sweep chunk. This is a requirement with batch marked TLHs for SATB. This is problematic for sweep because sweep must determine last obj in a chunk. With a batch marked TLH spanning multiple chunks, first slice of TLH contained in a chunk will be completely marked, all mark bits will be set to the end of the chunk, as a result the last live obj in chunk is undiscoverable. Hence, we must align the TLHs to sweep chunk boundaries when a pool freeEntry spans multiple chunks during TLH allocation.
To do so, we first determine the chunk boundaries containing the
freeEntry baseAddr. If freeEntry topAddr is past the chunk top boundary
(the freeEntry spans multiple chunks) then we clip part of the freeEntry
which flows into next chunk.

This method of aligning chunks is sensitive to specific heap resizing
that takes places
- After TLH allocation/alignment
- Before actual sweep and assignment of chunks.

When we're aligning TLHs we are essentially predicting the chunk
assignment during sweep, with heap resizing these predictions are wrong.
During alignment of TLH we expect different boundaries. This is only
problematic  for 1 specific case, LOA expansion during pre-collect. This
specific case shifts the boundaries of the pools such that the TLH that
is at the bounds of SOA/LOA ends up spanning multiple chunks (LOA
consumes allocated SOA).  To address this issue we preserve the old LOA
boundary for sweep purposes when we have this problematic case.

Signed-off-by: Salman Rana <salman.rana@ibm.com>