Core Implementation for SATB with Gencon
Introduce core SATB implementation required for generational collection
_(OpenJ9 gencon policy)_. Currently, SATB works with the standard
Concurrent Global Collector, only with a single generation configuration
_(optavgpause)_. To extend its functionality to generational collection,
we must limit SATB to the old/tenure generation and account for local
collector mutations.

- TLH alignment isn’t required for generational.  Make explicit check to
skip setting alignment params in MM_MemoryPoolAddressOrderedList  if we
have generational config _(scavengerEnabled)_. Furthermore, tlh
allocation query to the collector _(reservedForGCAllocCacheSize)_ must
return 0 in case of generational, we don’t require the tlh to be
reserved.
- Non-tlh allocations must be individually marked only if allocated
directly in tenure
- Introduce new global collector API: _objectTenured_, which is used by
external GC (Scavenger) when an object is tenured. Default
implementation in base class _ConcurrentGC_ is left empty, we only
require this for the SATB collector. objectTenured calls a new delegate
to handle the tenured object (call barrier on it).
- Local collector must notify Global collector of objet tenure
(MM_Scavenger::copy)
- Update OMR GC consts for downstream OpenJ9 barrier handling

Signed-off-by: Salman Rana <salman.rana@ibm.com>